Renamed EditableController to EditingController
